Forgot Your Password Page
Made it so that when clicking "FalconBooking" during login it will take you back to the main page. Also added in a new page in which you can reset your password in the event they forget it. Removed the sign-up feature as accounts will be given and cannot be signed up.