Added in Valid Email Checking
I added in a method to check if the email being used is a valid one, as well as added in a confirmation page. When the user logs in if the email or password is invalid then they can't login. If it is then they get to log in and then it redirects them to the confirmation page which then lets them go back to the main page