Completed all login functionalities
I added in valid email checking for forgotYourPassword (basically the same way Alex had done it for the login page but for this page). I also made the conformation page change depending on where the user previously was on. So, if the user had just logged in, the conformation page would display a your all set message and a button returning home. If the user had just reset their password, the conformation page would display different text and a button that goes back to the login page.